Fixes screensaver timeout setting
The idleTimeout setting does not seem to be located in the
com.apple.screensaver plist instead it is managed by one or more plist files
under the ByHost subdirectory.  This fix attempts to correctly define the
domain we need to write to.

When you update the domain, it must be 'defaults write ByHost/com.apple...';
don't even think of using the absolute pathname (i.e. 'defaults write
/Users/.../com.apple...' won't work)

This commit was inspired by @hungryzi
https://github.com/pivotal-sprout/sprout/pull/197